Start unification of QuakeGX (#33)
* Start unification of QuakeGX

* silence warning

* Unify draw header

* Add mathlib and pr_cmds to unified list

* add view and wad to unified files

* Fix rumble not being multiplayer safe on Wii

* host.c unified

* various merge cleanups

* Remove unnecessary ifdefs

* properly document ifdef

* console.c unified

* update paths and folder structure for wii

* protection against some crashes on wii

* Change scale to accept floats

* merge cl_hud.c

* Cleaup cl_hud merger and fix an input bug

* Upload GX specific files

* Remove duplicate file

* Upload Makefile.wii and fix ctr/psp compile errors

* Update Makefile to use the same build folder as other systems

* Removed references to SetDlightColor and fixed ifdefs

* Clean up host.c and quakedef.h

* Clean up sv_main.c

* Cleanup and document view.c

* Cleanup wad.c

+++ source/ctr/gl/gl_draw.c
@@ -519,43 +519,43 @@ void Draw_String (int x, int y, char *str)
 	Draw_ColoredString(x, y, str, 255, 255, 255, 255, 1); 
 }
 
-void Draw_ColoredString(int x, int y, char *str, float r, float g, float b, float a, int scale) 
+void Draw_ColoredString(int x, int y, char *str, float r, float g, float b, float a, float scale) 
 {
 	while (*str)
 	{
-		Draw_CharacterRGBA (x, y, *str, r, g, b, a, scale);
+		Draw_CharacterRGBA (x, y, *str, r, g, b, a, (int)scale);
 
 		// Hooray for variable-spacing!
 		if (*str == ' ')
-			x += 4 * scale;
+			x += 4 * (int)scale;
         else if ((int)*str < 33 || (int)*str > 126)
-            x += 8 * scale;
+            x += 8 * (int)scale;
         else
-            x += (font_kerningamount[(int)(*str - 33)] + 1) * scale;
+            x += (font_kerningamount[(int)(*str - 33)] + 1) * (int)scale;
 		
 		str++;
 	}
 }
 
-int getTextWidth(char *str, int scale)
+int getTextWidth(char *str, float scale)
 {
 	int width = 0;
 
     for (int i = 0; i < strlen(str); i++) {
         // Hooray for variable-spacing!
 		if (str[i] == ' ')
-			width += 4 * scale;
+			width += 4 * (int)scale;
         else if ((int)str[i] < 33 || (int)str[i] > 126)
-            width += 8 * scale;
+            width += 8 * (int)scale;
         else
-            width += (font_kerningamount[(int)(str[i] - 33)] + 1) * scale;
+            width += (font_kerningamount[(int)(str[i] - 33)] + 1) * (int)scale;
     }
 
 	return width;
 }
 
 
-void Draw_ColoredStringCentered(int y, char *str, float r, float g, float b, float a, int scale)
+void Draw_ColoredStringCentered(int y, char *str, float r, float g, float b, float a, float scale)
 {
 	Draw_ColoredString((vid.width - getTextWidth(str, scale))/2, y, str, r, g, b, a, scale);
 }
